requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
diskcache==5.6.3
filelock==3.24.3
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
psutil==7.2.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==14.3.3
shellingham==1.5.4
six==1.17.0
tqdm==4.67.3
typer==0.24.0
typer-slim==0.24.0
typing_extensions==4.15.0
